+++ src/main/java/kostuchenkov/rgr/repository/ProductRepository.java
@@ -1,7 +1,7 @@
 package kostuchenkov.rgr.repository;
 
-import kostuchenkov.rgr.domain.Product;
+import kostuchenkov.rgr.domain.product.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<Product, Integer> {
+public interface ProductRepository extends CrudRepository<Product, Long> {
 }
