+++ src/main/java/kostuchenkov/rgr/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package kostuchenkov.rgr.repository;
 
-import kostuchenkov.rgr.domain.User;
+import kostuchenkov.rgr.domain.user.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 }
